requirements/base.txt
Django<1.6
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2
johnny-cache==1.6
django-crispy-forms==1.4.0
django-allauth==0.18.0
django-messages==0.5.1
django-pipeline==1.3.26